Struct libsignal_service::content::DataMessage

pub struct DataMessage {
    pub body: Option<String>,
    pub attachments: Vec<AttachmentPointer>,
    pub group: Option<GroupContext>,
    pub group_v2: Option<GroupContextV2>,
    pub flags: Option<u32>,
    pub expire_timer: Option<u32>,
    pub profile_key: Option<Vec<u8>>,
    pub timestamp: Option<u64>,
    pub quote: Option<Quote>,
    pub contact: Vec<Contact>,
    pub preview: Vec<Preview>,
    pub sticker: Option<Sticker>,
    pub required_protocol_version: Option<u32>,
    pub is_view_once: Option<bool>,
    pub reaction: Option<Reaction>,
    pub delete: Option<Delete>,
    pub body_ranges: Vec<BodyRange>,
}

Fields
body: Option<String>attachments: Vec<AttachmentPointer>group: Option<GroupContext>group_v2: Option<GroupContextV2>flags: Option<u32>expire_timer: Option<u32>profile_key: Option<Vec<u8>>timestamp: Option<u64>quote: Option<Quote>contact: Vec<Contact>preview: Vec<Preview>sticker: Option<Sticker>required_protocol_version: Option<u32>is_view_once: Option<bool>reaction: Option<Reaction>delete: Option<Delete>body_ranges: Vec<BodyRange>
Implementations

impl DataMessage

pub fn body(&self) -> &str

Returns the value of body, or the default value if body is unset.

pub fn flags(&self) -> u32

Returns the value of flags, or the default value if flags is unset.

pub fn expire_timer(&self) -> u32

Returns the value of expire_timer, or the default value if expire_timer is unset.

pub fn profile_key(&self) -> &[u8]

Returns the value of profile_key, or the default value if profile_key is unset.

pub fn timestamp(&self) -> u64

Returns the value of timestamp, or the default value if timestamp is unset.

pub fn required_protocol_version(&self) -> u32

Returns the value of required_protocol_version, or the default value if required_protocol_version is unset.

pub fn is_view_once(&self) -> bool

Returns the value of is_view_once, or the default value if is_view_once is unset.